+++ Classes/Service/MailService.php
@@ -3,6 +3,7 @@ namespace T3v\T3vCore\Service;
 
 use TYPO3\CMS\Core\Mail\MailMessage;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Utility\MailUtility;
 
 use T3v\T3vCore\Service\AbstractService;
 
@@ -15,15 +16,16 @@ class MailService extends AbstractService {
   /**
    * Sends a mail.
    *
-   * @param array $from The sender
-   * @param array $to The recipient(s)
-   * @param array $cc The recipient(s) who will be copied in on the message
+   * @param array $from The sender of the message, an array consisting of a `name` and `address` key / value pair, the `name` entry is optional and can be omitted
+   * @param array $to The recipient(s), an array consisting of `name` and `address` key / value pairs, the `name` entry is optional and can be omitted
+   * @param array $cc The recipient(s) who will be copied in on the message, an array consisting of `name` and `address` key / value pairs, the `name` entry is optional and can be omitted
+   * @param array $replyTo The optional reply-to addresses of the message, an array consisting of `name` and `address` key / value pairs, the `name` entry is optional and can be omitted
    * @param string $subject The subject
    * @param string $message The message
-   * @param string $format The optional format either `text/plain` or `text/html`, defaults to `text/plain`
+   * @param string $format The optional format either `text/html` or `text/plain`, defaults to `text/html`
    * @return boolean If the mail was sent
    */
-  public function send($from, $to, $cc, $subject, $message, $format = 'text/plain') {
+  public function send($from, $to, $cc, $replyTo, $subject, $message, $format = 'text/html') {
     $subject = (string) $subject;
     $message = (string) $message;
     $format  = (string) $format;
@@ -32,19 +34,23 @@ class MailService extends AbstractService {
 
     // === From ===
 
-    if ($from) {
+    if (!empty($from)) {
       $name    = trim($from['name']);
       $address = trim($from['address']);
 
       if ($address && GeneralUtility::validEmail($address)) {
         if (!empty($name)) {
-          $mail->setFrom(array($address => $name));
+          $mail->setFrom([$address => $name]);
         } else {
           $mail->setFrom($address);
         }
       } else {
         return false;
       }
+    } else {
+      $systemFrom = MailUtility::getSystemFrom();
+
+      $mail->setFrom($systemFrom);
     }
 
     // === To / Cc ===
@@ -52,11 +58,10 @@ class MailService extends AbstractService {
     $toRecipients = [];
     $ccRecipients = [];
 
-    if (is_array($to)) {
+    if (!empty($to)) {
       $recipients = [];
 
       foreach ($to as $recipient) {
-        $name    = trim($recipient['name']);
         $address = trim($recipient['address']);
 
         if ($address && GeneralUtility::validEmail($address)) {
@@ -69,15 +74,15 @@ class MailService extends AbstractService {
         $name      = trim($recipient['name']);
         $address   = trim($recipient['address']);
 
-        if ($i == 0) { // First recipient
-          // Add first recipient to `To`
+        if ($i == 0) { // The first recipient
+          // Add the first recipient to `To`
           if (!empty($name)) {
             $toRecipients[$address] = $name;
           } else {
             $toRecipients[] = $address;
           }
-        } else { // Other recipients
-          // Add other recipients to `Cc`
+        } else { // The other recipients
+          // Add the other recipients to `Cc`
           if (!empty($name)) {
             $ccRecipients[$address] = $name;
           } else {
@@ -87,13 +92,13 @@ class MailService extends AbstractService {
       }
     }
 
-    if (count($toRecipients)) {
+    if (!empty($toRecipients)) {
       $mail->setTo($toRecipients);
     } else {
       return false;
     }
 
-    if (is_array($cc)) {
+    if (!empty($cc)) {
       foreach ($cc as $recipient) {
         $name    = trim($recipient['name']);
         $address = trim($recipient['address']);
@@ -108,9 +113,24 @@ class MailService extends AbstractService {
       }
     }
 
-    if (count($ccRecipients)) {
+    if (!empty($ccRecipients)) {
       $mail->setCc($ccRecipients);
     }
+
+    // === Reply To ===
+
+    // if (!empty($replyTo)) {
+    //   $name    = trim($replyTo['name']);
+    //   $address = trim($replyTo['address']);
+    //
+    //   if ($address && GeneralUtility::validEmail($address)) {
+    //     if (!empty($name)) {
+    //       $mail->setReplyTo($address, $name);
+    //     } else {
+    //       $mail->setReplyTo($address);
+    //     }
+    //   }
+    // }
 
     // === Subject ===